django_vue/django_complete/requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.3.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==24.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.10.0
tzdata==2024.1
